Avoid loading unnecessary components and images durating initail rendering
